+++ tests/Attacks/Attack3Test.php
@@ -28,7 +28,6 @@ class Attack3Test extends TestCase
             'int'                 => 1,
             'float'               => 1.1,
             'bool'                => true,
-            'null'                => null,
             'array'               => [1, 2, 3],
             'emptyArray'          => [],
             'union'               => 'string',