+++ tests/ValueObjects/VO.php
@@ -22,7 +22,6 @@ readonly class VO extends ValueObject
     public int $int;
     public float $float;
     public bool $bool;
-    public null $null;
     public array $array;
     public array $emptyArray;
     public string|int|float|bool|array $union;